Move testutil under /test and /test/extra
