+++ python/test/extra/feature_extractor_extratest.py
@@ -1,9 +1,12 @@
 import unittest
+
 from vmaf.config import VmafConfig, VmafExternalConfig
 from vmaf.core.asset import Asset
 from vmaf.core.feature_extractor import VmafFeatureExtractor
 from vmaf.core.matlab_feature_extractor import StrredFeatureExtractor, SpEEDMatlabFeatureExtractor
 from vmaf.tools.stats import ListStats
+
+from testutil import set_default_576_324_videos_for_testing
 
 __copyright__ = "Copyright 2016-2018, Netflix, Inc."
 __license__ = "Apache, Version 2.0"
@@ -100,19 +103,7 @@ class ParallelMatlabFeatureExtractorTestNew(unittest.TestCase):
     def test_run_SpEED_matlab_fextractor(self):
 
         print 'test on running SpEED (Matlab) feature extractor...'
-        ref_path = VmafConfig.test_resource_path("yuv", "src01_hrc00_576x324.yuv")
-        dis_path = VmafConfig.test_resource_path("yuv", "src01_hrc01_576x324.yuv")
-        asset = Asset(dataset="test", content_id=0, asset_id=0,
-                      workdir_root=VmafConfig.workdir_path(),
-                      ref_path=ref_path,
-                      dis_path=dis_path,
-                      asset_dict={'width': 576, 'height': 324})
-
-        asset_original = Asset(dataset="test", content_id=0, asset_id=1,
-                               workdir_root=VmafConfig.workdir_path(),
-                               ref_path=ref_path,
-                               dis_path=ref_path,
-                               asset_dict={'width': 576, 'height': 324})
+        ref_path, dis_path, asset, asset_original = set_default_576_324_videos_for_testing()
 
         self.fextractor = SpEEDMatlabFeatureExtractor(
             [asset, asset_original],
