+++ python/test/extra/quality_runner_extratest.py
@@ -1,11 +1,12 @@
 import unittest
 
 from vmaf.core.matlab_quality_runner import SpEEDMatlabQualityRunner, StrredQualityRunner
-from vmaf.tools.testutil import set_default_576_324_videos_for_testing
 from vmaf.config import VmafConfig, VmafExternalConfig
 from vmaf.core.asset import Asset
 from vmaf.core.quality_runner import PsnrQualityRunner, VmafQualityRunner, VmafossExecQualityRunner
 from vmaf.core.result_store import FileSystemResultStore
+
+from testutil import set_default_576_324_videos_for_testing
 
 __copyright__ = "Copyright 2016-2018, Netflix, Inc."
 __license__ = "Apache, Version 2.0"
