+++ python/test/feature_extractor_test.py
@@ -10,7 +10,8 @@ from vmaf.core.feature_extractor import VmafFeatureExtractor, MomentFeatureExtra
     PsnrFeatureExtractor, SsimFeatureExtractor, MsSsimFeatureExtractor, VifFrameDifferenceFeatureExtractor
 from vmaf.core.asset import Asset
 from vmaf.core.result_store import FileSystemResultStore
-from vmaf.tools.testutil import set_default_576_324_videos_for_testing, set_default_flat_1920_1080_videos_for_testing
+
+from testutil import set_default_576_324_videos_for_testing, set_default_flat_1920_1080_videos_for_testing
 
 
 class FeatureExtractorTest(unittest.TestCase):
